Added Script, Target, ID components. Changed Camera to Script component, loadable from tiled. Set the main character as a "target" property of a "camera" object and it'll set it up correctly.
Script: Any sort of gamelogic script. Usually isn't tied to anything else, but can be. Should be linked via a target
Target: Contains the ID of another object, typically for a script or movement reason (Script is what it effects, movement is "go here")
ID: Name, may contain other things later. Useful as it gives an easy way to reference things in tiled. The name should never be passed around as an identifier, only for tiled purposes. Heck it could probably be deleted after the level is loaded...or perhaps just never updated.